requirements.txt
python-dotenv==1.2.1
streamlit==1.51.0
PyPDF2==3.0.1
faiss-cpu==1.12.0

langchain==1.0.3
langchain-core==1.0.2
langchain-community==0.4.1
langchain-openai==1.0.1
langchain-text-splitters==1.0.0

openai>=1.51.0
